Move tree node to separate module
